Revert "[WordPress Plugin Directory] Request: WP Headless Menu Editor"
This reverts commit 1eae6858062b5b42ec530cd30c3f8ca2807ee666.

+++ includes/classes/menu_editor.class.php
@@ -1,17 +1,17 @@
 <?php
 
-// Create _MENU_EDITOR namespace
-namespace _MENU_EDITOR;
+// Create HME_MENU_EDITOR namespace
+namespace HME_MENU_EDITOR;
 
 // Import Models
-require PLUGIN_DIR_PATH . '/includes/models/menu-item.model.php';
-require PLUGIN_DIR_PATH . '/includes/models/post-menu-item.model.php';
+require HME_PLUGIN_DIR_PATH . '/includes/models/menu-item.model.php';
+require HME_PLUGIN_DIR_PATH . '/includes/models/post-menu-item.model.php';
 
-// use _MODELS namespace
-use _MODELS as _models;
+// use HME_MODELS namespace
+use HME_MODELS as hme_models;
 
-// MenuEditor Class
-class MenuEditor
+// HMEMenuEditor Class
+class HMEMenuEditor
 {
 
     public function __construct()
@@ -21,12 +21,12 @@ class MenuEditor
     public function init_ajax()
     {
         $ajax_functions = array(
-            'get_menus',
-            'get_menu_data',
-            'save_menu',
-            'get_posts',
-            'create_new_menu',
-            'delete_menu'
+            'hme_get_menus',
+            'hme_get_menu_data',
+            'hme_save_menu',
+            'hme_get_posts',
+            'hme_create_new_menu',
+            'hme_delete_menu'
         );
 
         foreach ($ajax_functions as $func) {
@@ -37,16 +37,16 @@ class MenuEditor
     }
 
     // Get initial data
-    public function get_initial_data()
+    public function hme_get_initial_data()
     {
-        $menus = $this->get_menus();
+        $menus = $this->hme_get_menus();
 
         if (count($menus) === 0) {
-            $this->create_menu(
+            $this->hme_create_menu(
                 'Main Menu',
-                $this->get_posts(['Page'])
+                $this->hme_get_posts(['Page'])
             );
-            $menus = $this->get_menus();
+            $menus = $this->hme_get_menus();
         }
 
         $locale = get_locale();
@@ -54,24 +54,24 @@ class MenuEditor
         return (object) array(
             "locale" => array("current" => $locale),
             "urls" => array(
-                "rest" => $this->get_rest_url() . 'hme/v1/',
-                "wpajax" => $this->get_ajax_url(),
-                "admin" => $this->get_admin_url()
+                "rest" => $this->hme_get_rest_url() . 'hme/v1/',
+                "wpajax" => $this->hme_get_ajax_url(),
+                "admin" => $this->hme_get_admin_url()
             ),
             "menus" => $menus,
-            "posts_types" => $this->get_posts_types()
+            "posts_types" => $this->hme_get_posts_types()
         );
     }
 
     // Create new menu
-    public function ajax_create_new_menu()
+    public function ajax_hme_create_new_menu()
     {
         $menu_name = $_POST["menu_name"];
         if (isset($menu_name)) {
-            $new_menu_id = $this->create_menu($menu_name, []);
+            $new_menu_id = $this->hme_create_menu($menu_name, []);
             $res = array(
                 'new_menu_id' => $new_menu_id,
-                'menus' => $this->get_menus()
+                'menus' => $this->hme_get_menus()
             );
             echo json_encode($res);
             wp_die();
@@ -79,14 +79,14 @@ class MenuEditor
     }
 
     // Delete menu
-    public function ajax_delete_menu()
+    public function ajax_hme_delete_menu()
     {
         $menu_id = $_POST['menu_id'];
-        if (isset($menu_id)) {
+        if (isset($menu_id) && intval($menu_id)) {
             wp_delete_nav_menu($menu_id);
             $res = array(
                 'deleted_menu_id' => $menu_id,
-                'menus' => $this->get_menus()
+                'menus' => $this->hme_get_menus()
             );
             echo json_encode($res);
             wp_die();
@@ -94,33 +94,33 @@ class MenuEditor
     }
 
     // Get WP REST API URL
-    public function get_rest_url()
+    public function hme_get_rest_url()
     {
         return str_replace('wp-admin', 'wp-json', admin_url());
     }
 
     // Get WP Ajax URL
-    public function get_ajax_url()
+    public function hme_get_ajax_url()
     {
         return admin_url('admin-ajax.php');
     }
 
     // Get WP Admin URL
-    public function get_admin_url()
+    public function hme_get_admin_url()
     {
         return admin_url();
     }
 
     // Create new menu
-    public function create_menu($menu_name, $menu_data)
+    public function hme_create_menu($menu_name, $menu_data)
     {
         $menu_id = wp_create_nav_menu($menu_name);
-        $this->save_menu($menu_id, $menu_data);
+        $this->hme_save_menu($menu_id, $menu_data);
         return $menu_id;
     }
 
     // Get menus
-    public function get_menus()
+    public function hme_get_menus()
     {
         return get_terms('nav_menu', array(
             'hide_empty' => false,
@@ -128,17 +128,17 @@ class MenuEditor
     }
 
     // AjaxService get menus
-    public function ajax_get_menus()
+    public function ajax_hme_get_menus()
     {
-        echo json_encode($this->get_menus());
+        echo json_encode($this->hme_get_menus());
         wp_die();
     }
 
     // AjaxService get menu items
-    public function ajax_get_menu_data()
+    public function ajax_hme_get_menu_data()
     {
-        if (isset($_GET["menu_id"])) {
-            $menu_id = $_GET["menu_id"];
+        $menu_id = $_GET["menu_id"];
+        if (isset($menu_id) && intval($menu_id)) {
             $menu_items = wp_get_nav_menu_items($menu_id);
             echo json_encode($menu_items);
             wp_die();
@@ -146,7 +146,7 @@ class MenuEditor
     }
 
     // Get posts types
-    public function get_posts_types()
+    public function hme_get_posts_types()
     {
         $args = array(
             'public' => true,
@@ -172,22 +172,26 @@ class MenuEditor
     }
 
     // AjaxService get posts
-    public function ajax_get_posts()
+    public function ajax_hme_get_posts()
     {
-        if (isset($_POST["types"])) {
+        $types = $_POST["types"];
+        if (isset($types)) {
 
             $types = $_POST["types"];
             $types = stripslashes($types);
             $types = json_decode($types, true);
-            $posts = $this->get_posts($types);
+
+            if (is_array($types)) {
+                $posts = $this->hme_get_posts($types);
 
             echo json_encode($posts);
             wp_die();
+            }
         }
     }
 
     // Get posts
-    private function get_posts($types)
+    private function hme_get_posts($types)
     {
         $res = [];
 
@@ -212,34 +216,36 @@ class MenuEditor
             $post->post_type_label = $post_object->labels->singular_name;
             $post->description = $post_object->description;
 
-            $post_menu_item = new _models\PostMenuItem($post);
+            $post_menu_item = new hme_models\PostMenuItem($post);
             array_push($res, $post_menu_item->$model);
         }
         return $res;
     }
 
     // AjaxService save menu
-    public function ajax_save_menu()
+    public function ajax_hme_save_menu()
     {
+        $menu_id = $_POST["menu_id"];
+        $menu_data = $_POST["menu_data"];
 
-        if (isset($_POST["menu_id"]) && isset($_POST["menu_data"])) {
-            $menu_id = $_POST["menu_id"];
-            $menu_data = $_POST["menu_data"];
+        if (isset($menu_id) && intval($menu_id) && isset($menu_data)) {
             $menu_data = stripslashes($menu_data);
             $menu_data = json_decode($menu_data, true);
 
-            $this->save_menu($menu_id, $menu_data);
-            echo json_encode(wp_get_nav_menu_items($menu_id));
-            wp_die();
-        };
+            if (is_array($menu_data)) {
+                $this->hme_save_menu($menu_id, $menu_data);
+                echo json_encode(wp_get_nav_menu_items($menu_id));
+                wp_die();
+            }
+        }
     }
 
     // Save menu
-    private function save_menu($menu_id, $menu_data)
+    private function hme_save_menu($menu_id, $menu_data)
     {
         try {
             foreach ($menu_data as $data) {
-                $menu_item = new _models\MenuItem($data);
+                $menu_item = new hme_models\MenuItem($data);
                 wp_update_nav_menu_item(
                     $menu_id,
                     $data["db_id"],
